backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
langgraph==0.3.21
langchain-openai==0.3.9
python-dotenv==1.1.0
